proper python package
